Move tests to correct PSR4 namespace
