+++ tests/App/Model/Parameter/Probe/PingParametersTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\App\Model\ProbeArgument;
+namespace App\Tests\App\Model\Parameter\Probe;
 
 use App\Model\Parameter\Probe\PingParameters;
 use PHPUnit\Framework\TestCase;
