+++ tests/App/Model/Parameter/Probe/TracerouteParametersTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\App\Model\ProbeArgument;
+namespace App\Tests\App\Model\Parameter\Probe;
 
 use App\Model\Parameter\NullParameters;
 use PHPUnit\Framework\TestCase;
 
-class NullParametersTest extends TestCase
+class TracerouteParametersTest extends TestCase
 {
     public function testDefault()
     {
